% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnsaf_functions.R
\name{dnsaf}
\alias{dnsaf}
\title{Calculate dNSAF}
\usage{
dnsaf(percolatorFile, fasta, test_run = F, q_cutoff = 0.01)
}
\arguments{
\item{percolatorFile}{a tibble read from the percolator target psms file}

\item{fasta}{a Biostrings AA string set containing a proteomics database used in the search}

\item{test_run}{Whether to do a short run of only 100 search results? Defaults to FALSE}

\item{q_cutoff}{Percolator q vsalue cutoff to consider as confident peptide. Defaults to 0.01}

\item{nonUniqueDummyCount}{ Dummy counts to give to a protein without unique peptide. Defaults to 0.5}
}
\description{
This function calculates distributed normalized spectral abundance factor from a percolator target psms file
}
\examples{
dnsaf(percolatorFile=percolator_target_psms)


}
\keyword{dNSAF}
\keyword{proteomics}
